+++ Api/Data/PaymentInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Zero1\OpenPos\Api\Data;
+
+/**
+ * Interface PaymentInterface
+ *
+ * @api
+ */
+interface PaymentInterface
+{
+}